move the contents of skeleton/interrupts into skeleton/init_trilogy
