+++ meadowphysics/main.c
@@ -18,7 +18,6 @@
 #include "i2c.h"
 #include "init_trilogy.h"
 #include "init_common.h"
-#include "interrupts.h"
 #include "monome.h"
 #include "timers.h"
 #include "adc.h"
