+++ whitewhale/config.mk
@@ -68,7 +68,6 @@ CSRCS = \
        ../skeleton/i2c.c     \
        ../skeleton/init_trilogy.c \
        ../skeleton/init_common.c \
-       ../skeleton/interrupts.c \
        ../skeleton/monome.c \
        ../skeleton/timers.c \
        ../skeleton/usb.c \
